légère simplification

+++ include/gpu_utils.cuh
@@ -5,6 +5,7 @@
 #include <cuda_runtime.h>
 
 // Fonction pour allouer la mémoire GPU
+//extern "C" void allocateBoidDataOnGPU(Types::BoidData& boids);
 void allocateBoidDataOnGPU(Types::BoidData& boids);
 
 // Fonction pour libérer la mémoire GPU
@@ -22,7 +23,7 @@ void reallocateIfNecessary(Types::BoidData& boids);
 // kernel
 __global__ void updateBoidsKernel(
     float* positionsX, float* positionsY, float* orientations, Types::Interaction* interactions,
-    const int numBoids, const float envWidth, const float envHeight, const float speed, const float angVelocity, const float timeStep,
+    const int numBoids, const int envWidth, const int envHeight, const float speed, const float angVelocity, const float timeStep,
     const float halvedFov, const float rDistancingSquared, const float rAlignmentSquared, const float rCohesionSquared,
     const float weightDistancing, const float weightAlignment, const float weightCohesion
 );
@@ -30,7 +31,7 @@ __global__ void updateBoidsKernel(
 // Fonction d'encapsulation pour appeler le kernel
 void updateBoidsCUDA(
     float* positionsX, float* positionsY, float* orientations, Types::Interaction* interactions,
-    const int numBoids, const float envWidth, const float envHeight, const float speed, const float angVelocity, const float timeStep,
+    const int numBoids, const int envWidth, const int envHeight, const float speed, const float angVelocity, const float timeStep,
     const float halvedFov, const float rDistancingSquared, const float rAlignmentSquared, const float rCohesionSquared,
     const float weightDistancing, const float weightAlignment, const float weightCohesion
 );
